# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Data Template Hub API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface RegisterInput {
  email: string;
  password: string;
  name: string;
}

export interface User {
  id: number;
  email: string;
  name: string;
  role: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  bio?: string | null;
  createdAt: string;
}

export interface AuthResult {
  user: User;
  token: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  icon: string;
  description: string;
  /** @nullable */
  color?: string | null;
  templateCount: number;
}

export interface Tag {
  id: number;
  name: string;
  count: number;
}

export interface Template {
  id: number;
  title: string;
  description: string;
  categoryId: number;
  categoryName: string;
  categorySlug?: string;
  authorId: number;
  authorName: string;
  /** @nullable */
  authorAvatarUrl?: string | null;
  version: string;
  status: string;
  downloads: number;
  averageRating: number;
  ratingCount: number;
  isFavorited: boolean;
  tags: string[];
  /** @nullable */
  language?: string | null;
  /** @nullable */
  complexity?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface TemplateDetail {
  id: number;
  title: string;
  description: string;
  categoryId: number;
  categoryName: string;
  categorySlug?: string;
  authorId: number;
  authorName: string;
  /** @nullable */
  authorAvatarUrl?: string | null;
  version: string;
  status: string;
  downloads: number;
  averageRating: number;
  ratingCount: number;
  isFavorited: boolean;
  tags: string[];
  /** @nullable */
  language?: string | null;
  /** @nullable */
  complexity?: string | null;
  code: string;
  documentation: string;
  /** @nullable */
  usageExample?: string | null;
  /** @nullable */
  prerequisites?: string | null;
  /** @nullable */
  changelog?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface TemplateContent {
  code: string;
  filename: string;
}

export interface TemplateListResult {
  templates: Template[];
  total: number;
  page: number;
  limit: number;
}

export interface TemplateInput {
  title: string;
  description: string;
  categoryId: number;
  code: string;
  documentation?: string;
  usageExample?: string;
  prerequisites?: string;
  tags: string[];
  version: string;
  language?: string;
  complexity?: string;
}

export interface TemplateUpdate {
  title?: string;
  description?: string;
  categoryId?: number;
  code?: string;
  documentation?: string;
  tags?: string[];
  version?: string;
  status?: string;
  language?: string;
  complexity?: string;
}

export interface TemplateVersion {
  id: number;
  templateId: number;
  version: string;
  changelog: string;
  code?: string;
  createdAt: string;
  authorName: string;
}

export interface Comment {
  id: number;
  templateId: number;
  authorId: number;
  authorName: string;
  /** @nullable */
  authorAvatarUrl?: string | null;
  content: string;
  createdAt: string;
}

export interface CommentInput {
  content: string;
}

export interface RatingInput {
  score: number;
}

export interface RatingResult {
  averageRating: number;
  ratingCount: number;
  userScore: number;
}

export interface Submission {
  id: number;
  title: string;
  description: string;
  categoryId: number;
  categoryName?: string;
  code?: string;
  documentation?: string;
  tags?: string[];
  status: string;
  submittedById: number;
  submittedByName: string;
  /** @nullable */
  reviewedById?: number | null;
  /** @nullable */
  reviewedByName?: string | null;
  /** @nullable */
  reviewNotes?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export interface SubmissionInput {
  title: string;
  description: string;
  categoryId: number;
  code: string;
  documentation?: string;
  tags: string[];
  language?: string;
  complexity?: string;
}

export interface RejectionInput {
  reason: string;
}

export interface AnalyticsSummary {
  totalTemplates: number;
  totalDownloads: number;
  totalUsers: number;
  totalCategories: number;
  pendingSubmissions: number;
  totalRatings: number;
}

export interface CategoryStat {
  categoryId: number;
  categoryName: string;
  categorySlug: string;
  templateCount: number;
  downloads: number;
}

export interface ContributorStat {
  userId: number;
  name: string;
  /** @nullable */
  avatarUrl?: string | null;
  templateCount: number;
  totalDownloads: number;
}

export interface TrendPoint {
  date: string;
  downloads: number;
}

export interface RoleUpdate {
  role: string;
}

export type ListTemplatesParams = {
/**
 * @nullable
 */
categoryId?: number | null;
/**
 * @nullable
 */
search?: string | null;
/**
 * @nullable
 */
tags?: string | null;
/**
 * @nullable
 */
sortBy?: string | null;
/**
 * @nullable
 */
page?: number | null;
/**
 * @nullable
 */
limit?: number | null;
};

export type ListSubmissionsParams = {
/**
 * @nullable
 */
status?: string | null;
};