# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Data Template Hub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
...
  /templates/{id}/comments:
    get:
      operationId: listTemplateComments
      tags: [comments]
      summary: Get comments for a template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"

    post:
      operationId: createComment
      tags: [comments]
      summary: Post a comment on a template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Comment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
...
  /favorites/{templateId}:
    post:
      operationId: addFavorite
      tags: [favorites]
      summary: Add template to favorites
      parameters:
        - name: templateId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Added to favorites

    delete:
      operationId: removeFavorite
      tags: [favorites]
      summary: Remove template from favorites
      parameters:
        - name: templateId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removed from favorites
  /ratings/{templateId}:
    post:
      operationId: rateTemplate
      tags: [ratings]
      summary: Rate a template (1-5)
      parameters:
        - name: templateId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RatingInput"
      responses:
        "200":
          description: Rating saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RatingResult"
  /submissions:
    get:
      operationId: listSubmissions
      tags: [submissions]
      summary: List template submissions (admin view)
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Submissions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Submission"

    post:
      operationId: submitTemplate
      tags: [submissions]
      summary: Submit a template for review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmissionInput"
      responses:
        "201":
          description: Submission created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
  /submissions/{id}/approve:
    post:
      operationId: approveSubmission
      tags: [submissions]
      summary: Approve a template submission (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Submission approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
  /submissions/{id}/reject:
    post:
      operationId: rejectSubmission
      tags: [submissions]
      summary: Reject a template submission (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectionInput"
      responses:
        "200":
          description: Submission rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
...
  /admin/users/{id}/role:
    patch:
      operationId: updateUserRole
      tags: [admin]
      summary: Update user role (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleUpdate"
      responses:
        "200":
          description: Role updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    RegisterInput:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
    AuthResult:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, email, name, role, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Comment:
      type: object
      required: [id, templateId, authorId, authorName, content, createdAt]
      properties:
        id:
          type: integer
        templateId:
          type: integer
        authorId:
          type: integer
        authorName:
          type: string
        authorAvatarUrl:
          type: ["string", "null"]
        content:
          type: string
        createdAt:
          type: string
    CommentInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
    RatingInput:
      type: object
      required: [score]
      properties:
        score:
          type: integer
    RatingResult:
      type: object
      required: [averageRating, ratingCount, userScore]
      properties:
        averageRating:
          type: number
        ratingCount:
          type: integer
        userScore:
          type: integer
    Submission:
      type: object
      required: [id, title, description, categoryId, status, submittedById, submittedByName, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        categoryId:
          type: integer
        categoryName:
          type: string
        code:
          type: string
        documentation:
          type: string
        tags:
          type: array
          items:
            type: string
        status:
          type: string
        submittedById:
          type: integer
        submittedByName:
          type: string
        reviewedById:
          type: ["integer", "null"]
        reviewedByName:
          type: ["string", "null"]
        reviewNotes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    SubmissionInput:
      type: object
      required: [title, description, categoryId, code, tags]
      properties:
        title:
          type: string
        description:
          type: string
        categoryId:
          type: integer
        code:
          type: string
        documentation:
          type: string
        tags:
          type: array
          items:
            type: string
        language:
          type: string
        complexity:
          type: string
    RejectionInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
...
    RoleUpdate:
      type: object
      required: [role]
      properties:
        role:
          type: string